Change SikuliX to be based on KasmVNC.
This makes a much larger image but a lot easier to use for common
scenario. Instead of requiring to access/share the host X11(/Wayland)
it runs a virtual X11 session.

Comes with a browser and scrcpy to remote control Android phones
(Waydroid does not run inside containers currently; or at least
it's hard to get it to run so it's an advanced option people could
still use).